+++ src/eca/features/prompt.clj
@@ -3,6 +3,7 @@
    [babashka.fs :as fs]
    [clojure.java.io :as io]
    [clojure.string :as string]
+   [eca.cache :as cache]
    [eca.features.skills :as f.skills]
    [eca.features.tools.mcp :as f.mcp]
    [eca.features.tools.util :as tools.util]
@@ -102,7 +103,7 @@
      :osName (str (System/getProperty "os.name") " " (System/getProperty "os.version"))
      :shell (or (System/getenv "SHELL") (System/getenv "ComSpec"))
      :userName (System/getProperty "user.name")
-     :homeDir (System/getProperty "user.home")
+     :homeDir (cache/user-home)
      :isSubagent (boolean (get-in db [:chats chat-id :subagent]))}
     (reduce
      (fn [m tool]